+++ CASAL2/source/Processes/Length/RecruitmentConstant.cpp
@@ -29,7 +29,7 @@ RecruitmentConstant::RecruitmentConstant(Model* model)
     partition_(model) {
   parameters_.Bind<string>(PARAM_CATEGORIES, &category_labels_, "Categories", "");
   parameters_.Bind<Double>(PARAM_PROPORTIONS, &proportions_, "Proportions", "", true);
-  parameters_.Bind<unsigned>(PARAM_LENGTH_BINS, &length_bins_, "The length bins that recruits are uniformly distributed over at the time of recruitment", "");
+  parameters_.Bind<double>(PARAM_LENGTH_BINS, &length_bins_, "The length bins that recruits are uniformly distributed over at the time of recruitment", "");
   parameters_.Bind<Double>(PARAM_R0, &r0_, "R0", "")->set_lower_bound(0.0);
 
   RegisterAsAddressable(PARAM_R0, &r0_);
@@ -64,8 +64,8 @@ void RecruitmentConstant::DoValidate() {
   if (proportions_.size() > 0) {
     if (proportions_.size() != category_labels_.size()) {
       LOG_ERROR_P(PARAM_PROPORTIONS)
-          << ": the number of proportions provided is not the same as the number of categories provided. Expected: "
-          << category_labels_.size()<< ", parsed " << proportions_.size();
+        << ": the number of proportions provided is not the same as the number of categories provided. Expected: "
+        << category_labels_.size()<< ", parsed " << proportions_.size();
     }
 
     Double proportion_total = 0.0;
@@ -74,7 +74,7 @@ void RecruitmentConstant::DoValidate() {
 
     if (!utilities::doublecompare::IsOne(proportion_total)) {
       LOG_WARNING() << parameters_.location(PARAM_PROPORTIONS)
-          <<": the proportions do not sum to 1.0. The proportions sum to " << AS_DOUBLE(proportion_total) << ". Auto-scaling the proportions to sum to 1.0";
+        <<": the proportions do not sum to 1.0. The proportions sum to " << AS_DOUBLE(proportion_total) << ". Auto-scaling the proportions to sum to 1.0";
 
       for (Double& proportion : proportions_)
         proportion /= proportion_total;
@@ -109,20 +109,20 @@ void RecruitmentConstant::DoExecute() {
   //Calculate new proportion totals to account for dynamic categories
   Double total_proportions = 0.0;
   for (auto category : partition_)
-   total_proportions += proportions_categories_[category->name_];
-
+    total_proportions += proportions_categories_[category->name_];
 
   //Update our partition with new recruitment values
   for (auto category : partition_) {
-   if (category->data_.size() != model_->length_bins().size())
-     LOG_CODE_ERROR() << "This function was written when categories were forced to have the same length bins as models. This is no longer the case.";
-   r0_by_length_bin_ = (r0_ * (proportions_categories_[category->name_]) / total_proportions) / length_bins_.size();
-   unsigned length_index = 0;
-   for (auto length_bin : length_bins_) {
-     length_index = std::distance(std::find(model_->length_bins().begin(), model_->length_bins().end(), length_bin), model_->length_bins().begin());
-     LOG_FINEST() << "putting " << r0_by_length_bin_ << " in category " << category->name_ << " in length bin " << model_->length_bins()[length_index];
-     category->data_[length_index] += r0_by_length_bin_;
-   }
+    if (category->data_.size() != model_->length_bins().size())
+      LOG_CODE_ERROR() << "This function was written when categories were forced to have the same length bins as models. This is no longer the case.";
+
+    r0_by_length_bin_ = (r0_ * (proportions_categories_[category->name_]) / total_proportions) / length_bins_.size();
+    unsigned length_index = 0;
+    for (auto length_bin : length_bins_) {
+      length_index = std::distance(std::find(model_->length_bins().begin(), model_->length_bins().end(), length_bin), model_->length_bins().begin());
+      LOG_FINEST() << "putting " << r0_by_length_bin_ << " in category " << category->name_ << " in length bin " << model_->length_bins()[length_index];
+      category->data_[length_index] += r0_by_length_bin_;
+    }
   }
 }
